requirements.txt
requests>=2.24.0
websocket-client>=1.4.1
